+++ lib/ui/screens/printers/edit/components/macro_group_list.dart
@@ -237,8 +237,8 @@ class _MacroGroup extends HookConsumerWidget {
 
   Icon? _macroAvatar(GCodeMacro macro) {
     if (!macro.visible) return const Icon(Icons.visibility_off_outlined);
-    if (!macro.showWhilePrinting) return const Icon(Icons.disabled_visible_outlined);
-    return null;
+    // if (!macro.s) return const Icon(Icons.disabled_visible_outlined);
+    return const Icon(Icons.visibility_outlined);
   }
 }